+++ packages/common/evmUtils/src/operations/transaction/getTransactionOperation.test.ts
@@ -14,20 +14,17 @@ describe('getTransactionOperation', () => {
 
     const request: Required<GetTransactionRequest> = {
       chain: EvmChain.create(chain, core),
-      subdomain: 'test.com',
       transactionHash: '0x9857d679ab331210161427d36d08c3b00e6d28c03366e9b891832ad9b5d478f7z',
     };
 
     const serializedRequest = getTransactionOperation.serializeRequest(request, core);
 
     expect(serializedRequest.chain).toBe(chain);
-    expect(serializedRequest.subdomain).toBe(request.subdomain);
     expect(serializedRequest.transactionHash).toBe(request.transactionHash);
 
     const deserializedRequest = getTransactionOperation.deserializeRequest(serializedRequest, core);
 
     expect((deserializedRequest.chain as EvmChain).apiHex).toBe(chain);
-    expect(deserializedRequest.subdomain).toBe(request.subdomain);
     expect(deserializedRequest.transactionHash).toBe(request.transactionHash);
   });
 });