+++ packages/common/streamsUtils/src/dataTypes/EvmStreamResult/EvmStreamResult.mock.ts
@@ -22,6 +22,8 @@ export const initialTest: EvmStreamResultInput = {
   nftTransfers: [],
   tag: '',
   streamId: '',
+  nftTokenApprovals: [],
+  nativeBalances: [],
 };
 
 const nativeTransaction: EvmStreamResultInput = {
@@ -67,6 +69,8 @@ const nativeTransaction: EvmStreamResultInput = {
     ERC721: [],
   },
   nftTransfers: [],
+  nftTokenApprovals: [],
+  nativeBalances: [],
 };
 
 const erc1155SingleTransfer: EvmStreamResultInput = {
@@ -154,6 +158,8 @@ const erc1155SingleTransfer: EvmStreamResultInput = {
       tokenSymbol: '',
     },
   ],
+  nftTokenApprovals: [],
+  nativeBalances: [],
 };
 
 const erc721Transfer: EvmStreamResultInput = {
@@ -323,6 +329,8 @@ const erc721Transfer: EvmStreamResultInput = {
       tokenContractType: 'ERC721',
     },
   ],
+  nftTokenApprovals: [],
+  nativeBalances: [],
 };
 
 const erc20LinkTransfer: EvmStreamResultInput = {
@@ -394,6 +402,8 @@ const erc20LinkTransfer: EvmStreamResultInput = {
     ERC721: [],
   },
   nftTransfers: [],
+  nftTokenApprovals: [],
+  nativeBalances: [],
 };
 
 const erc20USDCTransfers: EvmStreamResultInput = {
@@ -663,6 +673,8 @@ const erc20USDCTransfers: EvmStreamResultInput = {
     ERC721: [],
   },
   nftTransfers: [],
+  nftTokenApprovals: [],
+  nativeBalances: [],
 };
 
 const wethEvents: EvmStreamResultInput = {
@@ -2071,6 +2083,8 @@ const wethEvents: EvmStreamResultInput = {
     ERC721: [],
   },
   nftTransfers: [],
+  nftTokenApprovals: [],
+  nativeBalances: [],
 };
 
 const erc721Approval: EvmStreamResultInput = {
@@ -2194,6 +2208,8 @@ const erc721Approval: EvmStreamResultInput = {
     ],
   },
   nftTransfers: [],
+  nftTokenApprovals: [],
+  nativeBalances: [],
 };
 
 const erc1155Approval: EvmStreamResultInput = {
@@ -2360,6 +2376,65 @@ const erc1155Approval: EvmStreamResultInput = {
     ERC721: [],
   },
   nftTransfers: [],
+  nftTokenApprovals: [],
+  nativeBalances: [],
+};
+
+const nativeBalances = {
+  confirmed: true,
+  chainId: '0x1',
+  abi: [],
+  streamId: 'v900a834-a542-43f9-98f6-0f76caf65394',
+  tag: 'someString',
+  retries: 0,
+  block: {
+    number: '16140655',
+    hash: '0xfc0b1ae5d04b3c5a780f1f72658edc9e26a6340df4b1118a713652b6d043039a',
+    timestamp: '1670510543',
+  },
+  logs: [],
+  txs: [
+    {
+      hash: '0xd304ef0dd0df001d8f37de569af63585dbf7637247b17dc12224828960b4d830',
+      gas: '120000',
+      gasPrice: '17605149814',
+      nonce: '43904',
+      input:
+        '0xa9059cbb00000000000000000000000072f54455fbdbff1b8975ab2df599ff9f9afb5310000000000000000000000000000000000000000000000000000000000606a9e0',
+      transactionIndex: '16',
+      fromAddress: '0x0b5c4a7fcda49e0a8661419bb55b86161a86db2a',
+      toAddress: '0xdac17f958d2ee523a2206206994597c13d831ec7',
+      value: '0',
+      type: '2',
+      v: '0',
+      r: '52200684288032529078283440758240219132630446778024109656659293182067276426999',
+      s: '25805856329951682196861199677610079726197925451457581628939426805653323455347',
+      receiptCumulativeGasUsed: '1214527',
+      receiptGasUsed: '63209',
+      receiptContractAddress: null,
+      receiptRoot: null,
+      receiptStatus: '1',
+    },
+  ],
+  txsInternal: [],
+  erc20Transfers: [],
+  erc20Approvals: [],
+  nftTokenApprovals: [],
+  nftApprovals: {
+    ERC721: [],
+    ERC1155: [],
+  },
+  nftTransfers: [],
+  nativeBalances: [
+    {
+      address: '0x0b5c4a7fcda49e0a8661419bb55b86161a86db2a',
+      balance: '3921415148849464186',
+    },
+    {
+      address: '0xdac17f958d2ee523a2206206994597c13d831ec7',
+      balance: '2135786324',
+    },
+  ],
 };
 
 export const mockEvmStreamResult = {
@@ -2372,4 +2447,5 @@ export const mockEvmStreamResult = {
   ERC721_APPROVAL: erc721Approval,
   ERC1155_APPROVAL: erc1155Approval,
   WETH_EVENTS: wethEvents,
+  NATIVE_BALANCES: nativeBalances,
 };