+++ packages/evmApi/integration/mocks/endpoints/getDateToBlock.ts
@@ -7,7 +7,6 @@ export const mockGetDateToBlock = MockScenarios.create(
     getParams: ({ req }) => ({
       chain: req.url.searchParams.get('chain'),
       date: req.url.searchParams.get('date'),
-      providerUrl: req.url.searchParams.get('providerUrl'),
     }),
     url: '/dateToBlock',
   },