+++ packages/streams/src/mapping/txs-processor/TxsProcessor.test.ts
@@ -108,6 +108,8 @@ const batch: IWebhook = {
   nftTransfers: [],
   tag: 'Transaction',
   streamId: 'ba3b3c52-3dd3-4eb7-a2b7-4b61d3439c5e',
+  nativeBalances: [],
+  nftTokenApprovals: [],
 };
 
 describe('TxsProcessor', () => {