+++ packages/client/solApi/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @moralisweb3/client-sol-api
 
+## 2.10.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/client-api-utils@2.10.1
+  - @moralisweb3/common-core@2.10.1
+  - @moralisweb3/common-sol-utils@2.10.1
+
 ## 2.10.0
 
 ### Patch Changes