+++ packages/common/authUtils/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @moralisweb3/common-auth-utils
 
+## 2.10.1
+
+### Patch Changes
+
+- Updated dependencies [[`33230c43c`](https://github.com/MoralisWeb3/Moralis-JS-SDK/commit/33230c43c4a100f400fb862718e21ef487ca656c)]:
+  - @moralisweb3/common-evm-utils@2.10.1
+  - @moralisweb3/common-core@2.10.1
+  - @moralisweb3/common-sol-utils@2.10.1
+
 ## 2.10.0
 
 ### Patch Changes